update tools.js
The breadcrumb helper object now references the main navigation structure to determine the items in the loaf. If The navigation structure is not multi-layered this may provoke errors. Also, none of the menu items can be duplicated as this may refer to later parent categories in the navigation. 

Future update: when clicking on the parent element in the breadcrumb, this would bring up the drop-down used in the navigation bar. This could be useful if the breadcrumb is referenced after the main content.