Update tools.js
When you are validating the forms it makes it easier to control which buttons are going to be disabled if any on submit. With this when you instantiate the method with .send() you can add the id of the submit button which you would like to be disabled after submit and not disabled if not submitted. This gives the user another chance if they get things wrong.