Breadcrumb Generator - Single Page No Parent
If a page does not have a parent page this needs to be addressed as well. This will now work however, there are two ways that this set up the breadcrumb one based on the navigation and another based on the URL structure. This can get confusing and may cause further issues. In future updates it may be prudent to add the ability of the user to select which of the two they prefer in either or all contexts.